data-assistant/flask-backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sympy==1.13.1
tenacity==9.0.0
torch==2.6.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
